requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
pillow==10.4.0
pyyaml==6.0.2
pydantic==2.9.2
python-dotenv==1.0.1
click==8.1.7
requests==2.32.3
jinja2==3.1.4
python-multipart==0.0.12
httpx==0.27.2
numpy

# Optional: Azure Blob Storage backend (required only when STORAGE_BACKEND=azure_blob)
# Uncomment and install: pip install azure-storage-blob
azure-storage-blob==12.23.1
